+++ src/server/common/CMakeLists.txt
@@ -139,17 +139,17 @@ cxx_test(common_HierarchyTest
          gtest_main
         )
 
-add_library(common_HNodeDsl
-            HNodeDsl.h
-            HNodeDsl.cpp
+add_library(common_HNodeGroup
+            HNodeGroup.h
+            HNodeGroup.cpp
            )
-target_link_libraries(common_HNodeDsl
+target_link_libraries(common_HNodeGroup
                       common_Hierarchy
                       common_string
                      )
-cxx_test(common_HNodeDslTest
-         HNodeDslTest.cpp
-         common_HNodeDsl
+cxx_test(common_HNodeGroupTest
+         HNodeGroupTest.cpp
+         common_HNodeGroup
          gtest_main
         )
