+++ src/server/common/HNodeGroup.h
